feat: Adding permissions to views and controllers

+++ app/Http/Controllers/GroupController.php
@@ -10,6 +10,8 @@ class GroupController extends Controller
 {
     public function index()
     {
+        $this->authorize('index', Group::class);
+
         $groups = Group::with('permissions')->get();
 
         $permissions = Permission::with([
@@ -22,11 +24,15 @@ class GroupController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Group::class);
+
         return view('groups.create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', Group::class);
+
         $request->validate([
             'name' => 'required|string|max:255',
         ]);
@@ -38,11 +44,15 @@ class GroupController extends Controller
 
     public function edit(Group $group)
     {
+        $this->authorize('update', Group::class);
+
         return view('groups.edit', compact('group'));
     }
 
     public function update(Request $request, Group $group)
     {
+        $this->authorize('update', Group::class);
+
         $request->validate([
             'name' => 'required|string|max:255',
         ]);
@@ -54,6 +64,8 @@ class GroupController extends Controller
 
     public function destroy(Group $group)
     {
+        $this->authorize('delete', Group::class);
+
         $group->delete();
 
         return redirect()->route('groups.index')->with('success', 'Group deleted successfully.');
